internal: share response timing helpers
